Allow passing in multiple filter queries
This is needed because sometimes a caller needs to be able to pass
multiple fq values. For example, I haven't found a way to use frange
with an unrelated boolean query. The following throws a parse error:

    +myfield:"valA" {!frange l=2}termfreq(myfield,"valB")

But if expressed as two separate filters, it behaves as expected:

    +myfield:"valA"
    {!frange l=2}termfreq(myfield,"valB")

This could be considered a breaking change. Queries will continue
working as before, but the caching behavior will be different for
calls that passed in a vector as a value for :fq. See the examples in
the docs[0] for how the caching will change.

Callers who are currently passing a vector can restore the previous
behavior by calling `(clojure.string/join " " <the vector>)`.

[0] https://solr.apache.org/guide/8_11/common-query-parameters.html#fq-filter-query-parameter